Version 1.1.0 Update.
Added support for plugins to 'connect' to the Bridge.

This support is only useful for those who wish to shade the project into their plugins.

Also added Faction Strikes to the IFaction class. (addStrike, removeStrike, getTotalStrikes and clearStrikes have been added).

Also added an asString method for IFactionPlayer.

BridgeAlreadyConnectedException - called when a plugin tries to connect to the FactionsBridge to set it up when it is already set up.